Rename FormyTest1 class and refactor: implement Constructor and Dispose (shared setup/cleanup code without sharing object instances)
xUnit.net creates a new instance of the test class for every test that is run, so any code which is placed into the constructor of the test class will be run for every single test.
See https://xunit.net/docs/shared-context
Note: if a test encounters an exception, Dispose() method will quit the browser driver.